Added slideshow capability
Use F1-F10 to start slideshow with 1-10 second delay.  It will wait at
least that long, or as long as it takes to play the longest gif
currently shown at least once.

There's still a crash I ran into going from 8 to 1 and back to 8.  More
testing needed.

Also slideshows do not automatically go fullscreen.  ESC now exits
fullscreen and slideshow mode and only if neither was on does it
quit the program.