Broken commit, saving for the heck of it
So from the SDL2 wiki here https://wiki.libsdl.org/CategoryRender

"This API is not designed to be used from multiple threads, see SDL bug #1995 for details."

Which I should have realized earlier since OpenGL has the same issues.

So I need to reorganize and separate loading the image pixels from
creating the the SDL_Textures and only do the former (which is the
slowest part anyway) on another thread.