Working threading with HW acceleration
Want to try on Windows out of curiosity before switching to SDL
threads.

Also, Note, since the slideshow functionality just sends a RIGHT button
event, if any current image is zoomed in it'll pan instead of going to
the next image(s).  I'm not sure if this is a bug or correct behavior
but I'm leaning toward the latter since the user had to interact with
the mouse wheel or + key to zoom in, they can hit 'f' (for best fit)
to get the slide show to continue.