Minor fixups, tried some static analysis tools
I tried clang-tidy, cppcheck and splint and I think that's about the
ordering from best to worst.  cppcheck says there's a memory leak when I
know there's not and splint just spits out a *ton* of useless output,
warnings and things that aren't problems.  clang-tidy only said one
thing, a branch based on a potentially uninitialized variable and that
was a valid issue.