Starts up almost instantly even without a thread
In hindsight, I should have known stbi_info was the culprit since it has
to open/read a bit/close each file to determine if it's an image
stb_image can handle.  I should probably remove it from the -f/-u
options as well.

The issue is now the image list the directory listing (all regular files
anyway).  I suppose the variable is g->files, but it's wasted memory and
load_image will try every one as you traverse the list, even if it's
already tried them.  Sanitizing as you go in load_new_images is
non-trivial and more importantly, ugly.  I'm torn, and still trying to
think of a cleaner way to both startup quickly and have a valid list as
quickly as possible.