Change loading behavior
Because for a fast startup we now put *all* files in directory in the
list, we can no longer just jump by n_imgs when moving in multi image
mode because for for example if the first one had to skip 3 files to
find an image, the next image needs to start at +n_imgs+3.

To do this we now move based on the index in the top left or lower
right, for moving left and right respectively, and each successive image
loaded starts 1 to the left or right of the previous.

This is a change, and may seem weird if for example you individually
move images and then go back to no-selection mode and move.  It now
doesn't matter what the other 3 images are the next set of n will be the
next n contiguous images immediately after the bottom right or before the
top left image.

Needs more testing but seems to work.