Simplified startup for -f/-u
So we now just load all arguments into g->files as they are, urls,
non-images, whatever.  And now the everywhere load_image is called, if
it fails we try curl_image and if that succeeds, we replace the url
with the downloaded file path in g->files and load that.