Some minor progress on windows...
I hate working on windows.  So the program compiles and runs
(though there are a couple warnings due to windows being windows)
and everything seems to work.

But the installer stuff sucks.  I turned off automatically asking
for administrator privileges last time because there's no reason you
shouldn't be able to install it anywhere.  However, if you try to
install in the default Program Files it will fail and you have to abort
and restart with "Run as Administrator".  There's no way to *add*
administrator privileges halfway through.

Same with the uninstaller once it is installed, you have to right click,
run as admin or it won't do a thing (won't even say it failed it just
won't do anything).

Finally, window's file associations/open with are giving me trouble ever
since I installed it to my user directory once.  Even though I'm 4 or 5
uninstalls and reinstalls since then, always in Program Files, I cannot
set it as the default program for images, or even as an option.  I
select it and it just won't show up in the list.

Registry stuff sucks too but that's a whole separate issue I just added
because I wanted to support the Add/Remove programms method of
uninstalling.  Surprise, it doesn't work.