Windows still sucks
Had to rename DELETE because winnt.h defines a DELETE macro and I don't
even know how that's getting pulled in but it's not me directly.

Also got rid of some warnings about %lu for size_t.  clang supports %zu
but gcc on windows doesn't for some reason.  So I changed them to
PRIuMAX (so ugly vs zu) but switched to defaulting to clang anyway.

Then built and packaged and installed/tested 0.96.  Seems to work.