First pass at arbitrary rotation...
It works but it's slow on large gifs for obvious reasons and I currently
do not give the user the option to save them, even for static images.

It's a little buggy and this is definitely not the final
design/implementation.  I git rid of the 90 degree rotation as a
separate thing and just use the arbitrary degree code for both and only
update the pixels at the end if they choose to save it.

Also fixed a tiny memory leak and a "possible leak" according to
valgrind from the thread not returning before exit.