Update CVector to fix idiotic bug...
This is why using a project/library is still the best testing.  That bug
has been there for years (wasn't covered in the test suite) till I ran
into it in this project which prompted me to update CVector before
propagating the new version to all my projects that use it.