Redraw when switching thumb mode on/off
Thumb mode is not finished, almost all user interactions need to be
disabled when in thumb mode and those that don't have different results.
So if you're not a programmer and don't want to deal with potential
bugs, use the latest official release (0.97 as of this commit)

Need to add the ability to double click a thumb to jump to single
viewing that image.

It'd also be cool to load the thumbs in a thread (and generate any
missing ones) so that the thumbs would appear as they're generated.