Set scroll to show current image when switching to list mode
I spent too long trying to figure out the "correct" way to calculate it
but what I got is close enough.  I think there might be a bug in the
scrolling code in nuklear because you can scroll past the end a bit
which even if by design makes no sense and looks stupid.

And there's disagreement between list.total_height and the calculated "target".