Fix threading bug that led to rare crashes
It was really stupid but obviously I shouldn't be loading
new images before the main thread has finished handling
the last set of newly loaded images because that process
is accessing both sets of image structures (allocating
new textures and freeing the old images) before switching
to point at the new ones in my pseudo "double buffering"
setup.  If the loading thread tries to load during that
time, well obviously there's conflict leading to double
frees etc.

Fortunately all loads are initiated through try_move so
it was simple to add waiting for done_loading to be 0
(meaning the main thread is finished "harvesting" the
last set of loaded images and switching.