Fix some bugs I introduced when I ...
changed how I load/store gifs.  I don't store the per frame delays in
between the frames anymore.  There was really no point since I was using
the delay for the first frame for the whole thing anyway, so now I just
store that in the structure and the frames are packed together.  I just
forgot to update the flip and rotation functions.