First draft at saving/loading preferences
Need to think about cache directory and other possible preferences to
add.  The cache dir defaults to prefpath/cache/date but can be
overridden with a command line argument.  If the cachedir was saved
in the config file it would have to be prefpath/cache/ without the date
to make sense, and the date would be added on at runtime.  Similarly,
if the user put a different directory there, the date should also be
added at runtime (and cut off when writing out at exit).

Now we have a minor inconsistency.  The argument sets the cache
directory absolutely, whereas the loaded/saved directory is a parent
directory that has date subdirs.  Need to think about it.