Always add the start image to the list even if...
It's not found in the directory scan because it doesn't have one of the
extensions we're searching for.  Before this change, clicking on an image that
doesn't have an extension would do nothing because not finding the
starting image in the list was an error I deemed worthy of exiting,
probably because I figured it'd be to unexpected to try to open up
something the user things is an image and instead get all the other
images in the directory.