Fixed a couple bugs, working on another bug/feature
Fix infinite loop bug when none of current files are valid
Switch to File Selector to open new files instead.

Fixed a crash when deleting the first image in the list.

Right now we exit when removing/deleting all images in thumb mode,
trying to change it to switch to File Selector as well, but it might
not be worth it.