Fix incorrect title bugs
When switching between different image modes I wasn't
correctly updating the window title sometimes. Specifically
I was setting img_focus to NULL for any mode switch and
not updating the title to img[0] when img_focus wasn't
NULL before (and thus that image was the title).

Now I correctly preserve img_focus when I can (when going
to any higher N-mode or a lower one >= img_focus) and update
title to img[0] and img_focus to NULL for the other cases.