Switch to HW-accelerated renderer by default...
because of some minor bugs/artifacts in GUI rendering and better scaling
(at least on some platforms that I've observed).  Still need to verify
that it's an SDL2 bug that's causing the GUI difference between software
and hardware rendering, but I'm 99% sure.

Made it compile time configurable because there are reasons for using software
rendering.  It required some changes, a few more minor
ones in the GUI to go.