Fix crash on MacOS, SDL_SetWindowTitle()...
Apparently SDL_SetWindowTitle() is only supposed to be called on the
main thread. The SDL2 wiki doesn't mention this and it's never been an
issue on Linux and Windows, the only platforms I officially support
since I don't have a Mac, but they do say so in the SDL3 wiki.

https://wiki.libsdl.org/SDL2/SDL_SetWindowTitle

vs

https://wiki.libsdl.org/SDL3/SDL_SetWindowTitle

Pretty important thing to leave out of SDL2 docs if you ask me.